+++ app/src/API/Handlers/UARTHandler.cpp
@@ -472,7 +472,7 @@ API::Handlers::UARTHandler::getBaudRateList(const QString &id,
 
   QJsonArray baudRates;
   for (const auto &rate : baudRateList)
-    baudRates.append(rate);
+    baudRates.append(rate.toInt());
 
   QJsonObject result;
   result[QStringLiteral("baudRateList")] = baudRates;
